enhancement: integrar os dados com o Stripe, sendo Room como um produto

+++ Hotel.Domain/Handlers/RoomHandlers/HandleDeleteRoom.cs
@@ -13,6 +13,10 @@ public partial class RoomHandler
         if (room.Reservations.Count > 0)
             throw new InvalidOperationException("Não foi possível deletar a hospedagem pois tem reservas associadas a ela. Sugiro que desative a hospedagem.");
 
+        //Delete room in Stripe
+        if (await _stripeService.DisableProductAsync(room.StripeProductId) == null)
+            throw new Exception("Não foi possível desabilitar o produto no serviço do Stripe.");    
+
         _repository.Delete(room);
         await _repository.SaveChangesAsync();
         return new Response("Hospedagem deletada com sucesso!", new { id });
